curate: Set shared parser separately
This is necessary to register subparsers recursively.
