+++ augur/curate/abbreviate_authors.py
@@ -10,6 +10,7 @@ import re
 from typing import Generator, List
 from augur.io.print import print_err
 from augur.utils import first_line
+from .argparse_shared_parser import shared_parser
 
 
 def parse_authors(
@@ -52,7 +53,7 @@ def register_parser(
 ) -> argparse._SubParsersAction:
     parser = parent_subparsers.add_parser(
         "abbreviate-authors",
-        parents=[parent_subparsers.shared_parser],  # type: ignore
+        parents=[shared_parser],  # type: ignore
         help=first_line(__doc__),
     )
