+++ augur/curate/apply_geolocation_rules.py
@@ -5,6 +5,7 @@ from collections import defaultdict
 from augur.errors import AugurError
 from augur.io.print import print_err
 from augur.utils import first_line
+from .argparse_shared_parser import shared_parser
 
 
 class CyclicGeolocationRulesError(AugurError):
@@ -188,7 +189,7 @@ def transform_geolocations(geolocation_rules, geolocation):
 
 def register_parser(parent_subparsers):
     parser = parent_subparsers.add_parser("apply-geolocation-rules",
-        parents=[parent_subparsers.shared_parser],
+        parents=[shared_parser],
         help=first_line(__doc__))
 
     parser.add_argument("--region-field", default="region",
