+++ augur/curate/normalize_strings.py
@@ -7,11 +7,12 @@ in cases where strings contain diacritics (see https://unicode.org/faq/normaliza
 import unicodedata
 
 from augur.utils import first_line
+from .argparse_shared_parser import shared_parser
 
 
 def register_parser(parent_subparsers):
     parser = parent_subparsers.add_parser("normalize-strings",
-        parents=[parent_subparsers.shared_parser],
+        parents=[shared_parser],
         help=first_line(__doc__))
 
     optional = parser.add_argument_group(title="OPTIONAL")
