+++ augur/curate/parse_genbank_location.py
@@ -10,6 +10,7 @@ import argparse
 from typing import Generator, List
 from augur.io.print import print_err
 from augur.utils import first_line
+from .argparse_shared_parser import shared_parser
 
 
 def parse_location(
@@ -50,7 +51,7 @@ def register_parser(
 ) -> argparse._SubParsersAction:
     parser = parent_subparsers.add_parser(
         "parse-genbank-location",
-        parents=[parent_subparsers.shared_parser],  # type: ignore
+        parents=[shared_parser],  # type: ignore
         help=first_line(__doc__),
     )
