+++ augur/curate/passthru.py
@@ -2,11 +2,12 @@
 Pass through records without doing any data transformations.
 Useful for testing, troubleshooting, or just converting file formats.
 """
+from .argparse_shared_parser import shared_parser
 
 
 def register_parser(parent_subparsers):
     return parent_subparsers.add_parser("passthru",
-        parents=[parent_subparsers.shared_parser],
+        parents=[shared_parser],
         help=__doc__)
